trading-agents/market-analyst/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-cloud-aiplatform==1.38.0
